separate appointment page and profile page
