+++ .vscode/launch.json
@@ -6,6 +6,7 @@
     "configurations": [
         {
             "name": "Flutter",
+            "program": "lib/main.dart",
             "request": "launch",
             "type": "dart"
         }
