+++ lib/appointment.dart
@@ -1,6 +1,4 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/material.dart';
-import 'package:Tuter/users.dart';
 
 class Appointment {
   final String className;
