+++ lib/profile.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:Tuter/auth.dart';
+
+
+class ProfilePage extends StatefulWidget {
+  @override
+
+  _ProfilePage createState() => _ProfilePage();
+}
+
+class _ProfilePage extends State<ProfilePage> {
+
+  final Auth _auth = Auth();
+
+  @override
+
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Profile'),
+        actions: <Widget>[
+          FlatButton.icon(
+            icon: Icon(Icons.person),
+            label: Text('Log Out'),
+            onPressed: () async {
+              await _auth.logOut();
+            },
+          )
+        ],
+      ),
+    );
+  }
+}
